Separate auth and admin controllers
